wip: scanner
