+++ yas/src/info/const_info/starrail.rs
@@ -1,6 +1,6 @@
 use crate::info::starrail::StarRailWindowInfo;
-
-use super::WindowSize;
+use super::ui::WindowSize;
+// use super::*;
 
 pub fn get_window_info(size: WindowSize) -> &'static StarRailWindowInfo {
     match size {
