+++ yas/src/item/mod.rs
@@ -1,2 +1,7 @@
 pub mod genshin_artifact;
 pub mod starrail_relic;
+
+pub enum Item {
+    GenshinArtifact(genshin_artifact::GenshinArtifact),
+    StarrailRelic(starrail_relic::StarrailRelic),
+}
