+++ yas/src/scanner/genshin.rs
@@ -17,64 +17,18 @@ use tract_onnx::prelude::tract_itertools::Itertools;
 use crate::item::genshin_artifact::{
     ArtifactSetName, ArtifactSlot, ArtifactStat, GenshinArtifact,
 };
-use crate::capture;
 use crate::common::character_name::CHARACTER_NAMES;
 use crate::common::color::Color;
 #[cfg(target_os = "macos")]
 use crate::common::utils::get_pid_and_ui;
-use crate::common::{utils, Rect, PixelRectBound};
+use crate::common::*;
 use crate::inference::inference::CRNNModel;
 use crate::inference::pre_process::{pre_process, to_gray, ImageConvExt};
-use crate::info::info::ScanInfo;
+use crate::info::ScanInfo;
 
 // Playcover only, wine should not need this.
 #[cfg(all(target_os = "macos", target_arch = "aarch64"))]
 use crate::common::utils::mac_scroll;
-
-pub struct YasScanner {
-    model: Arc<CRNNModel>,
-    enigo: Enigo,
-
-    info: ScanInfo,
-    config: YasScannerConfig,
-
-    row: u32,
-    col: u32,
-
-    pool: f64,
-
-    initial_color: Color,
-
-    // for scrolls
-    scrolled_rows: u32,
-    avg_scroll_one_row: f64,
-
-    avg_switch_time: f64,
-    scanned_count: u32,
-
-    is_cloud: bool,
-}
-
-enum ScrollResult {
-    TimeLimitExceeded,
-    Interrupt,
-    Success,
-    Skip,
-}
-
-#[derive(Debug)]
-pub struct YasScanResult {
-    name: String,
-    main_stat_name: String,
-    main_stat_value: String,
-    sub_stat_1: String,
-    sub_stat_2: String,
-    sub_stat_3: String,
-    sub_stat_4: String,
-    level: String,
-    equip: String,
-    star: u32,
-}
 
 impl YasScanResult {
     pub fn to_internal_artifact(&self) -> Option<GenshinArtifact> {
@@ -140,35 +94,6 @@ fn calc_pool(row: &Vec<u8>) -> f32 {
     }
     // pool /= len as f64;
     pool
-}
-
-impl YasScanner {
-    pub fn new(info: ScanInfo, is_cloud: bool, model: &[u8], content: String) -> YasScanner {
-        let row = info.art_row;
-        let col = info.art_col;
-
-        let model = CRNNModel::new(model, content).expect("Err");
-
-        YasScanner {
-            enigo: Enigo::new(),
-            model: Arc::new(model),
-            info,
-            config: YasScannerConfig::parse(),
-
-            row,
-            col,
-
-            pool: -1.0,
-            initial_color: Color::default(),
-            scrolled_rows: 0,
-            avg_scroll_one_row: 0.0,
-
-            avg_switch_time: 0.0,
-            scanned_count: 0,
-
-            is_cloud,
-        }
-    }
 }
 
 impl YasScanner {
